+++ policy.go
@@ -1,4 +1,4 @@
-// Copyright 2013-2020 Aerospike, Inc.
+// Copyright 2014-2021 Aerospike, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -44,7 +44,10 @@ type BasePolicy struct {
 	FilterExpression *FilterExpression
 
 	// Priority of request relative to other transactions.
-	// Currently, only used for scans.
+	// Only used for scans on server versions < 4.9.
+	//
+	// Priority is obsolete and will eventually be removed.
+	// Use ScanPolicy.RecordsPerSecond instead of priority.
 	Priority Priority //= Priority.DEFAULT;
 
 	// ReadModeAP indicates read policy for AP (availability) namespaces.
@@ -92,7 +95,12 @@ type BasePolicy struct {
 	//
 	// Default for read: 2 (initial attempt + 2 retries = 3 attempts)
 	//
-	// Default for write/query/scan: 0 (no retries)
+	// Default for write: 0 (no retries)
+	//
+	// Default for partition scan or query with nil filter: 5
+	// (6 attempts. See ScanPolicy comments.)
+	//
+	// No default for legacy scan/query. No retries are allowed for these commands.
 	MaxRetries int //= 2;
 
 	// SleepBetweenRtries determines the duration to sleep between retries.  Enter zero to skip sleep.
@@ -125,12 +133,10 @@ type BasePolicy struct {
 	// The default is to not send the user defined key.
 	SendKey bool // = false
 
-	// UseCompression tells the server to compress its response using zlib.
-	// Use zlib compression on write or batch read commands when the command buffer size is greater
-	// than 128 bytes. In addition, it directs the server to compress its response on read commands.
-	// The server response compression threshold is also 128 bytes. Works with server EE v4.8.0.1+.
+	// UseCompression uses zlib compression on command buffers sent to the server and responses received
+	// from the server when the buffer size is greater than 128 bytes.
 	//
-	// This option will increase CPU and memory usage (for extra compressed buffers), but
+	// This option will increase cpu and memory usage (for extra compressed buffers),but
 	// decrease the size of data sent over the network.
 	//
 	// Default: false