+++ predexp.go
@@ -101,7 +101,7 @@ func (e *predExpAnd) String() string {
 }
 
 // NewPredExpAnd creates an AND predicate. Argument describes the number of expressions.
-func NewPredExpAnd(nexpr uint16) *predExpAnd {
+func NewPredExpAnd(nexpr uint16) PredExp {
 	return &predExpAnd{nexpr: nexpr}
 }
 
@@ -128,7 +128,7 @@ func (e *predExpOr) String() string {
 }
 
 // NewPredExpOr creates an OR predicate. Argument describes the number of expressions.
-func NewPredExpOr(nexpr uint16) *predExpOr {
+func NewPredExpOr(nexpr uint16) PredExp {
 	return &predExpOr{nexpr: nexpr}
 }
 
@@ -154,7 +154,7 @@ func (e *predExpNot) String() string {
 }
 
 // NewPredExpNot creates a NOT predicate
-func NewPredExpNot() *predExpNot {
+func NewPredExpNot() PredExp {
 	return &predExpNot{}
 }
 
@@ -180,7 +180,7 @@ func (e *predExpIntegerValue) String() string {
 }
 
 // NewPredExpIntegerValue embeds an int64 value in a predicate expression.
-func NewPredExpIntegerValue(val int64) *predExpIntegerValue {
+func NewPredExpIntegerValue(val int64) PredExp {
 	return &predExpIntegerValue{val: val}
 }
 
@@ -207,7 +207,7 @@ func (e *predExpStringValue) String() string {
 }
 
 // NewPredExpStringValue embeds a string value in a predicate expression.
-func NewPredExpStringValue(val string) *predExpStringValue {
+func NewPredExpStringValue(val string) PredExp {
 	return &predExpStringValue{val: val}
 }
 
@@ -234,7 +234,7 @@ func (e *predExpGeoJSONValue) String() string {
 }
 
 // NewPredExpGeoJSONValue embeds a GeoJSON value in a predicate expression.
-func NewPredExpGeoJSONValue(val string) *predExpGeoJSONValue {
+func NewPredExpGeoJSONValue(val string) PredExp {
 	return &predExpGeoJSONValue{val: val}
 }
 
@@ -268,32 +268,32 @@ func (e *predExpBin) String() string {
 }
 
 // NewPredExpUnknownBin creates a Bin predicate expression which its type is not known.
-func NewPredExpUnknownBin(name string) *predExpBin {
+func NewPredExpUnknownBin(name string) PredExp {
 	return &predExpBin{name: name, tag: _AS_PREDEXP_UNKNOWN_BIN}
 }
 
 // NewPredExpUnknownBin creates a Bin predicate expression which its type is integer.
-func NewPredExpIntegerBin(name string) *predExpBin {
+func NewPredExpIntegerBin(name string) PredExp {
 	return &predExpBin{name: name, tag: _AS_PREDEXP_INTEGER_BIN}
 }
 
 // NewPredExpUnknownBin creates a Bin predicate expression which its type is String.
-func NewPredExpStringBin(name string) *predExpBin {
+func NewPredExpStringBin(name string) PredExp {
 	return &predExpBin{name: name, tag: _AS_PREDEXP_STRING_BIN}
 }
 
 // NewPredExpUnknownBin creates a Bin predicate expression which its type is GeoJSON.
-func NewPredExpGeoJSONBin(name string) *predExpBin {
+func NewPredExpGeoJSONBin(name string) PredExp {
 	return &predExpBin{name: name, tag: _AS_PREDEXP_GEOJSON_BIN}
 }
 
 // NewPredExpUnknownBin creates a Bin predicate expression which its type is List.
-func NewPredExpListBin(name string) *predExpBin {
+func NewPredExpListBin(name string) PredExp {
 	return &predExpBin{name: name, tag: _AS_PREDEXP_LIST_BIN}
 }
 
 // NewPredExpUnknownBin creates a Bin predicate expression which its type is Map.
-func NewPredExpMapBin(name string) *predExpBin {
+func NewPredExpMapBin(name string) PredExp {
 	return &predExpBin{name: name, tag: _AS_PREDEXP_MAP_BIN}
 }
 
@@ -322,17 +322,17 @@ func (e *predExpVar) String() string {
 }
 
 // NewPredExpIntegerVar creates 64 bit integer variable used in list/map iterations.
-func NewPredExpIntegerVar(name string) *predExpVar {
+func NewPredExpIntegerVar(name string) PredExp {
 	return &predExpVar{name: name, tag: _AS_PREDEXP_INTEGER_VAR}
 }
 
 // NewPredExpStringVar creates string variable used in list/map iterations.
-func NewPredExpStringVar(name string) *predExpVar {
+func NewPredExpStringVar(name string) PredExp {
 	return &predExpVar{name: name, tag: _AS_PREDEXP_STRING_VAR}
 }
 
 // NewPredExpGeoJSONVar creates GeoJSON variable used in list/map iterations.
-func NewPredExpGeoJSONVar(name string) *predExpVar {
+func NewPredExpGeoJSONVar(name string) PredExp {
 	return &predExpVar{name: name, tag: _AS_PREDEXP_GEOJSON_VAR}
 }
 
@@ -379,17 +379,17 @@ func (e *predExpMD) marshal(cmd *baseCommand) error {
 }
 
 // NewPredExpRecDeviceSize creates record size on disk predicate
-func NewPredExpRecDeviceSize() *predExpMD {
+func NewPredExpRecDeviceSize() PredExp {
 	return &predExpMD{tag: _AS_PREDEXP_REC_DEVICE_SIZE}
 }
 
 // NewPredExpRecLastUpdate creates record last update predicate
-func NewPredExpRecLastUpdate() *predExpMD {
+func NewPredExpRecLastUpdate() PredExp {
 	return &predExpMD{tag: _AS_PREDEXP_REC_LAST_UPDATE}
 }
 
 // NewPredExpRecVoidTime creates record expiration time predicate expressed in nanoseconds since 1970-01-01 epoch as 64 bit integer.
-func NewPredExpRecVoidTime() *predExpMD {
+func NewPredExpRecVoidTime() PredExp {
 	return &predExpMD{tag: _AS_PREDEXP_REC_VOID_TIME}
 }
 
@@ -427,7 +427,7 @@ func (e *predExpMDDigestModulo) marshal(cmd *baseCommand) error {
 // 		NewPredExpIntegerValue(1),
 // 		NewPredExpIntegerEqual(),
 // )
-func NewPredExpRecDigestModulo(mod int32) *predExpMDDigestModulo {
+func NewPredExpRecDigestModulo(mod int32) PredExp {
 	return &predExpMDDigestModulo{mod: mod}
 }
 
@@ -474,52 +474,52 @@ func (e *predExpCompare) marshal(cmd *baseCommand) error {
 }
 
 // NewPredExpIntegerEqual creates Equal predicate for integer values
-func NewPredExpIntegerEqual() *predExpCompare {
+func NewPredExpIntegerEqual() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_INTEGER_EQUAL}
 }
 
 // NewPredExpIntegerUnequal creates NotEqual predicate for integer values
-func NewPredExpIntegerUnequal() *predExpCompare {
+func NewPredExpIntegerUnequal() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_INTEGER_UNEQUAL}
 }
 
 // NewPredExpIntegerGreater creates Greater Than predicate for integer values
-func NewPredExpIntegerGreater() *predExpCompare {
+func NewPredExpIntegerGreater() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_INTEGER_GREATER}
 }
 
 // NewPredExpIntegerGreaterEq creates Greater Than Or Equal predicate for integer values
-func NewPredExpIntegerGreaterEq() *predExpCompare {
+func NewPredExpIntegerGreaterEq() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_INTEGER_GREATEREQ}
 }
 
 // NewPredExpIntegerLess creates Less Than predicate for integer values
-func NewPredExpIntegerLess() *predExpCompare {
+func NewPredExpIntegerLess() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_INTEGER_LESS}
 }
 
 // NewPredExpIntegerLessEq creates Less Than Or Equal predicate for integer values
-func NewPredExpIntegerLessEq() *predExpCompare {
+func NewPredExpIntegerLessEq() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_INTEGER_LESSEQ}
 }
 
 // NewPredExpStringEqual creates Equal predicate for string values
-func NewPredExpStringEqual() *predExpCompare {
+func NewPredExpStringEqual() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_STRING_EQUAL}
 }
 
 // NewPredExpStringUnequal creates Not Equal predicate for string values
-func NewPredExpStringUnequal() *predExpCompare {
+func NewPredExpStringUnequal() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_STRING_UNEQUAL}
 }
 
 // NewPredExpGeoJSONWithin creates Within Region predicate for GeoJSON values
-func NewPredExpGeoJSONWithin() *predExpCompare {
+func NewPredExpGeoJSONWithin() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_GEOJSON_WITHIN}
 }
 
 // NewPredExpGeoJSONContains creates Region Contains predicate for GeoJSON values
-func NewPredExpGeoJSONContains() *predExpCompare {
+func NewPredExpGeoJSONContains() PredExp {
 	return &predExpCompare{tag: _AS_PREDEXP_GEOJSON_CONTAINS}
 }
 
@@ -536,7 +536,7 @@ func (e *predExpStringRegex) String() string {
 }
 
 // NewPredExpStringRegex creates a Regex predicate
-func NewPredExpStringRegex(cflags uint32) *predExpStringRegex {
+func NewPredExpStringRegex(cflags uint32) PredExp {
 	return &predExpStringRegex{cflags: cflags}
 }
 
@@ -579,32 +579,32 @@ func (e *predExpIter) String() string {
 }
 
 // NewPredExpListIterateOr creates an Or iterator predicate for list items
-func NewPredExpListIterateOr(name string) *predExpIter {
+func NewPredExpListIterateOr(name string) PredExp {
 	return &predExpIter{name: name, tag: _AS_PREDEXP_LIST_ITERATE_OR}
 }
 
 // NewPredExpMapKeyIterateOr creates an Or iterator predicate on map keys
-func NewPredExpMapKeyIterateOr(name string) *predExpIter {
+func NewPredExpMapKeyIterateOr(name string) PredExp {
 	return &predExpIter{name: name, tag: _AS_PREDEXP_MAPKEY_ITERATE_OR}
 }
 
 // NewPredExpMapValIterateOr creates an Or iterator predicate on map values
-func NewPredExpMapValIterateOr(name string) *predExpIter {
+func NewPredExpMapValIterateOr(name string) PredExp {
 	return &predExpIter{name: name, tag: _AS_PREDEXP_MAPVAL_ITERATE_OR}
 }
 
 // NewPredExpListIterateAnd creates an And iterator predicate for list items
-func NewPredExpListIterateAnd(name string) *predExpIter {
+func NewPredExpListIterateAnd(name string) PredExp {
 	return &predExpIter{name: name, tag: _AS_PREDEXP_LIST_ITERATE_AND}
 }
 
 // NewPredExpMapKeyIterateAnd creates an And iterator predicate on map keys
-func NewPredExpMapKeyIterateAnd(name string) *predExpIter {
+func NewPredExpMapKeyIterateAnd(name string) PredExp {
 	return &predExpIter{name: name, tag: _AS_PREDEXP_MAPKEY_ITERATE_AND}
 }
 
 // NewPredExpMapKeyIterateAnd creates an And iterator predicate on map values
-func NewPredExpMapValIterateAnd(name string) *predExpIter {
+func NewPredExpMapValIterateAnd(name string) PredExp {
 	return &predExpIter{name: name, tag: _AS_PREDEXP_MAPVAL_ITERATE_AND}
 }